economy cog added, queue commands added
update requirements to include pandas. pull economy into its own file. economy is now its own cog. created a cogs folder to hold all cogs
