+++ bot.py
@@ -4,7 +4,6 @@ import logging
 from discord.ext import commands
 import helper
 from get_assignments import get_new_assignments, datetime_file
-import blackjack
 
 #note for me:
 # when using python keyword in terminal, u must reference the direct path to the venv python executable.
@@ -31,7 +30,8 @@ bot.remove_command('help')
 # on the event called `on_ready`, python terminal shows that the bot is logged in by printing 
 @bot.event
 async def on_ready():
-    await bot.load_extension("blackjack")
+    await bot.load_extension("cogs.blackjack")
+    await bot.load_extension("cogs.economy")
     return print(f'I\'m logged in as {bot.user}')
 
 # defines help command. uses bot.group decorator to enable help to take further inputs after help - so that the
