+++ requirements.txt
@@ -11,7 +11,9 @@ discord.py==2.0.1
 frozenlist==1.3.1
 idna==3.4
 multidict==6.0.2
+numpy==1.24.2
 packaging==21.3
+pandas==1.5.3
 pydantic==1.10.2
 pyparsing==3.0.9
 python-dateutil==2.8.2
